security-platform-flask/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-login==0.6.3
flask-bcrypt==1.0.1
flask-socketio==5.3.6
flask-limiter==3.5.0
python-socketio==5.10.0
eventlet==0.34.2
python-dotenv==1.0.0
openai==1.6.0
marshmallow==3.20.1
flasgger==0.9.7.1
